rename ファイル構成をアーキテクチャに合わせる
