+++ lib/controller/weather_controller.dart
@@ -1,8 +1,8 @@
-import 'package:flutter_training/domain/weather_service.dart';
 import 'package:flutter_training/model/weather_model.dart';
+import 'package:flutter_training/service/weather_service.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-part 'weather_provider.g.dart';
+part 'weather_controller.g.dart';
 
 @Riverpod(keepAlive: true)
 WeatherService weatherService(WeatherServiceRef ref) {
